+++ test/complex-validations-test.js
@@ -19,21 +19,33 @@ buster.testCase('complex validations', {
             author: lewd.optional({ firstName: String, lastName: String }),
             chapters: lewd.all([{
                 name: String,
+                inProgress: lewd(Boolean).optional().default(false),
                 pages: lewd.all(lewd.integer(), lewd.range({ min: 1 }))
             }], lewd.len({ min: 3 }).because('a book needs at least three chapters (where: ${path})'))
         });
         
         buster.referee.refute.exception(function () {
-            condition({
+            var result = condition({
                 title: 'My Book',
                 chapters: [
                     { name: 'Chapter 1', pages: 4 },
-                    { name: 'Chapter 2', pages: 2 },
+                    { name: 'Chapter 2', pages: 2, inProgress: true },
                     { name: 'Chapter 3', pages: 11 },
                 ]
             });
+            
+            buster.referee.assert.same(result.chapters[0].inProgress, false);
+            buster.referee.assert.same(result.chapters[1].inProgress, true);
+            buster.referee.assert.same(result.chapters[2].inProgress, false);
         });
 
+        assertViolationAt(function () {
+            condition({
+                title: 'My Book With A Way Too Long Title',
+                chapters: []
+            });
+        }, ['title']);
+        
         assertViolationAt(function () {
             condition({
                 title: 'My Book',